Simplify how extra files by attributes are collect in incremental generation mode (#2967)
Having an `XcodeProjAutomaticTargetProcessingInfo.extra_files` makes it
clear that we are only looking at those attributes in order to collect
them. e.g. `hdrs` isn’t used for anything in incremental generation
mode, so it doesn’t make sense to have it be a named attribute.

This will also be a memory savings once we drop legacy generation mode.

Signed-off-by: Brentley Jones <github@brentleyjones.com>