Stop including test hosts in schemes when `adjust_schemes_for_swiftui_previews = True`
This can break some people’s workflows (because a test host is set to only compile for a certain target environment). Also, the test host is implicitly included in the scheme at the right action.

Other transitive dependencies are still built in the wrong action. In a future change we need to have test-based targets have their transitive dependencies listed under the test action, not run.

Signed-off-by: Brentley Jones <github@brentleyjones.com>