Add initial support for rules_swift’s `mixed_language_library`
This currently results in oddly named internal targets, because we can’t read the `SwiftInteropInfo` on the `_clang` target, and because we don’t perform target merging. By implementing target merging for `mixed_language_library` (in a future change) we will sidestep the `SwiftInteropInfo` issue (though it’s still an issue in general we will want to address).

Signed-off-by: Brentley Jones <github@brentleyjones.com>