Use only the debug settings from the most-downstream Swift target (#3073)
This prevents settings which are specific to a particular Swift module
compilation, such as clang module maps in a mixed-language library, from
getting accumulated with the public settings. Before this change using
`mixed_language_library` would result in module redefinition errors in
lldb.

I’m not sure if this has a chance of not setting some flags that would
need to be set. In local testing everything worked, but as with all
things lldb, wider user testing is needed. Ideally long term we can use
serialized debug info instead (and/or explicit modules).

**Note:** For this to work the best, the single dep of a top-level
target needs to be a `swift_library` or `mixed_language_target`. This
change might result in lldb regressions for custom setups that don't
meet that requirement.

---------

Signed-off-by: Brentley Jones <github@brentleyjones.com>