Expand "Make" variables inside of values in the env dictionary on test rules (#3088)
Fixes #3056

This PR implements a gap introduced when bazelbuild/rules_apple#2476 was
introduced, where expandable env settings on test rules would not be
expanded into the xcscheme during generation. There are a couple of
things that reviewers should be aware of before proceeding with this
change:

- This change is not considerate of the rules_apple being used, meaning
that it assumes the developer is using at least version
[3.6.0](https://github.com/bazelbuild/rules_apple/releases/tag/3.6.0)
(when the breaking changes in #2476 were introduced). This is not good
and should be considered a blocker, but I don't know how to address this
with the tools offered by the ruleset or Bazel.
- This change as-implemented requires regeneration of the xcschemes if
the configured value of the "Make" variable ever changes.

Signed-off-by: Aaron Sky <aaronsky@skyaaron.com>