Add `compile_only` aspect (#3156)
Adds an aspect that can be used to skip actions which are not
compilation related. This is useful for warming projects in CI to
maintain a Bazel remote cache for users to use. It is also useful when
compilation actions are the only thing that need to be validated as it
allows skipping potentially longer actions like bundling, signing, etc.

Example usage: `bazel run //:LocalDevCacheWarming_xcodeproj -- 'build
--aspects=//:compile_only_aspect.bzl%compile_only_aspect
--output_groups=compiles $_GENERATOR_LABEL_'`

Signed-off-by: Luis Padron <heyluispadron@gmail.com>