Skip targets without "compiles" output group in compile_only aspect (#3172)
We are running some additional targets which are not Apple / Swift
specific but still required and the build was failing since there was no
"compiles" output group.

I could add those rules to the list in the aspect definition. However I
fear that it would set the bad example and the list would probably grow
a lot over time.

Signed-off-by: Adin Cebic <cebic.ad@gmail.com>