Fix incorrect SHA2 for rules_swift (#3210)
The `rules_swift` release page shows the hash that is listed in this
file, but that hash is incorrect. The actual hash from downloading the
file, as well as what *GitHub* shows as the hash, is this one.

<img width="772" alt="Screenshot 2025-07-10 at 5 56 16 AM"
src="https://github.com/user-attachments/assets/aceb8f83-315c-4e08-b48a-296c65d1e62d"
/>

Signed-off-by: Prasanna Swaminathan <prasanna.swaminathan@gmail.com>