Fix Metal toolchains in Xcode 26 betas (#3213)
Fixes:

```
Error: TOOLCHAINS was set to 'com.apple.dt.toolchain.Metal.32023' but the default toolchain was found, that likely means a matching toolchain isn't installed
```

Apparently `$toolchain` either was never set for Metal toolchains in
earlier Xcode versions or Xcode didn't complain when it was set in
`action_env`. I tested on Xcode 16 and it was never there while it
definitely shows up in Xcode 26 Beta 3.

I think version check may not be needed here at all. If yes, please do
let me know what would be suggested way of performing the check. I tried
calling `xcodebuild -version` but that seem to randomly fail for some
people so I removed it.

Signed-off-by: Adin Cebic <cebic.ad@gmail.com>