Use DEVELOPER_DIR as an --xcode_version
In the exceptional case where the user has multiple versions of Xcode on
his system with the same Xcode version, we want to make sure that Xcode
is invoking bazel by referring to itself.

As bazel supports passing DEVELOPER_DIR to --xcode_version, we're using
that variable when invoked through Xcode instead.