Allow separate/merged index build output base to be configured (#3243)
We've run into stability issues that we believe can be traced back to
the point where the build and indexbuild output bases were merged,
particularly in the area of sudden pcm corruption when changing branches
that require a module cache busting. I'm investigating whether or not
this helps to ease the pain.

TODO:
- [x] docs
- [x] changelog

---------

Signed-off-by: Aaron Sky <aaronsky@skyaaron.com>