Support StoreKit Configuration setting in xcscheme Run action (#3245)
Fixes #3130.

Redoes #3182 (and potentially fixes unresolved comments on #3185). 

The StoreKit Testing configuration file has a peculiar requirement that
has no precedent in the rules. The file's path in the scheme must be
relative to the `xcshareddata/xcschemes` directory the scheme is under,
not the scheme itself (so one level shallower than you'd expect). Making
the path relative to the project fails, and making it absolute crashes
Xcode.

A prior iteration of this PR attempted to relativize the path in
Starlark, but exposed some safety shortcomings. In knowing the absolute
path of both the scheme directory and the configuration file, doing this
work in the generator gives us slightly improved safety guarantees at
the cost of performance. To this author, there is too much ambiguity
between the paths of either location without this information, since the
files can potentially exist in independent file trees under a workspace.

---------

Signed-off-by: Aaron Sky <aaronsky@skyaaron.com>