Adjust when we build for code coverage when it isn't enabled (#3257)
Fixes #3255 (partially)

For builds where code coverage is not enabled in the scheme's test
action, do not build with coverage. Should reduce thrash in index builds
and address a regression in the cache performance of Xcode builds that
hadn't opted into this feature.

Note: this does not fix the issue for build-for-running builds of tests,
as we have no clear-cut way of differentiating between build-for-running
and build-for-testing in the BazelDependencies run script phase. Open to
other ideas, since this is based on undocumented xcodebuild behavior and
is not super-nice.

Signed-off-by: Aaron Sky <aaronsky@skyaaron.com>